refactor(scene/render): Camera.CullingMask uint32 → uint64 + SetCullingMask(Layer)
- CullingMask type uint32_t → uint64_t (SP5 Task 3)
- SetCullingMask(Layer) / SetCullingMask(uint64_t) 2 overload 추가
- 기본값 Layer::All (= ~0ull) 명시 사용
- scene_renderer.cpp:236 의 0u → 0ull (uint64 비교 일관)
- CollectFromActor 시그니처 uint32_t → uint64_t (header + impl)
- 옛 직접 멤버 접근 (cam->CullingMask = ...) 은 자동 promotion

SP5 Task 3 완료.

+++ src/render/scene_renderer.h
@@ -41,7 +41,7 @@ namespace SJH
         /// @brief Actor 트리 DFS — MeshRenderer 수집 + Queue 에 Submit.
         /// @param cullingMask Camera::GetCullingMask() — actor.GetLayer() 와 AND 검사로 필터 (SP4 D-15).
         ///        자식 트리는 visibleToCamera 와 무관하게 계속 traverse (자식이 다른 layer 일 수 있음).
-        void CollectFromActor(const Scene::Actor& actor, const vmath::mat4& viewMat, uint32_t cullingMask);
+        void CollectFromActor(const Scene::Actor& actor, const vmath::mat4& viewMat, uint64_t cullingMask);
 
         /// @brief Actor 트리 DFS — IsActive + IsEnabled Camera 컴포넌트를 out 에 수집.
         void CollectCameras(const Scene::Actor& actor, std::vector<Scene::Camera*>& out);
